Optimize Direct3D-S2 for 6GB VRAM: Fix OOM, Add Extreme Offloading, Update Docs

+++ Direct3D-S2/direct3d_s2/models/refiner/unet3d.py
@@ -234,7 +234,7 @@ class ResnetBlock3D(nn.Module):
 
         hidden_states = input_tensor
         dtype = hidden_states.dtype
-        hidden_states = self.norm1(hidden_states.float()).to(dtype)
+        hidden_states = self.norm1(hidden_states)
         hidden_states = self.nonlinearity(hidden_states)
 
         if self.upsample is not None:
@@ -249,7 +249,7 @@ class ResnetBlock3D(nn.Module):
 
         hidden_states = self.conv1(hidden_states)
 
-        hidden_states = self.norm2(hidden_states.float()).to(dtype)
+        hidden_states = self.norm2(hidden_states)
 
         hidden_states = self.nonlinearity(hidden_states)
 
@@ -644,7 +644,7 @@ class UNet3DModel(nn.Module):
  
         if self.conv_norm_out:
             dtype = sample.dtype
-            sample = self.conv_norm_out(sample.float()).to(dtype)
+            sample = self.conv_norm_out(sample)
             sample = self.conv_act(sample)
         if self.conv_out!=None:
             sample = self.conv_out(sample)
